version 2 with gemini description/requirements.txt
Flask
ultralytics
opencv-python
requests
google-generativeai
numpy
